nvim: deletes image plugin
image plugin was introducing errors. It caused the `:messages` command
not to work, plus there was an error being logged on exit →
```
uv_close: Assertion '!uv__is_closing(handle)' failed
```
Both were fixed after removing this plugin. To be fair, the error might
be related to the imagemagick pgk or something else.